--- 1/src/main/java/com/whoisacat/edu/testingApp1/dao/MyCsvReaderException.java
@@ -1,8 +1,0 @@
-package com.whoisacat.edu.testingApp1.dao;
-
-public class MyCsvReaderException extends Exception{
-
-    public MyCsvReaderException(){
-        super("somethingWrongWithCvReader");
-    }
-}
