--- 1/src/main/java/com/whoisacat/edu/testingApp1/dao/QuestionDao.java
@@ -1,10 +1,0 @@
-package com.whoisacat.edu.testingApp1.dao;
-
-import com.whoisacat.edu.testingApp1.domain.Question;
-
-import java.util.List;
-
-public interface QuestionDao{
-
-    List<Question> loadObjectList() throws MyCsvReaderException;
-}
