--- 1/src/main/java/com/whoisacat/edu/testingApp1/service/QuestionReaderService.java
@@ -1,11 +1,0 @@
-package com.whoisacat.edu.testingApp1.service;
-
-import com.whoisacat.edu.testingApp1.dao.MyCsvReaderException;
-import com.whoisacat.edu.testingApp1.domain.Question;
-
-import java.util.List;
-
-public interface QuestionReaderService{
-
-    List<Question> readList() throws MyCsvReaderException;
-}
