--- 1/src/main/java/com/whoisacat/edu/testingApp1/service/QuestionWriterService.java
@@ -1,6 +1,0 @@
-package com.whoisacat.edu.testingApp1.service;
-
-public interface QuestionWriterService{
-
-    void writeQuestionsList();
-}
