--- 1/src/main/resources/questions.csv
@@ -1,6 +1,0 @@
-id,question
-1,What is Java mean
-2,What is Spring
-3,What is CSV
-4,Would you like a cup of tea
-5,What is different between meaning and knowledge
