+++ 2/src/main/java/com/whoisacat/edu/testingApp2/dao/QuestionDao.java
@@ -1,0 +1,10 @@
+package com.whoisacat.edu.testingApp2.dao;
+
+import com.whoisacat.edu.testingApp2.domain.Question;
+
+import java.util.List;
+
+public interface QuestionDao{
+
+    List<Question> loadObjectList() throws WHOCsvReaderException;
+}
