+++ 2/src/main/java/com/whoisacat/edu/testingApp2/dao/WHOCsvReaderException.java
@@ -1,0 +1,8 @@
+package com.whoisacat.edu.testingApp2.dao;
+
+public class WHOCsvReaderException extends Exception{
+
+    public WHOCsvReaderException(Exception e){
+        super("somethingWrongWithCvReader",e);
+    }
+}
