--- 2/test-app-2/src/main/java/com/whoisacat/edu/testingApp2/dao/MyCsvReaderException.java
@@ -1,8 +1,0 @@
-package com.whoisacat.edu.testingApp2.dao;
-
-public class MyCsvReaderException extends Exception{
-
-    public MyCsvReaderException(Exception e){
-        super("somethingWrongWithCvReader",e);
-    }
-}
