--- 2/test-app-2/src/main/java/com/whoisacat/edu/testingApp2/dao/QuestionDao.java
@@ -1,10 +1,0 @@
-package com.whoisacat.edu.testingApp2.dao;
-
-import com.whoisacat.edu.testingApp2.domain.Question;
-
-import java.util.List;
-
-public interface QuestionDao{
-
-    List<Question> loadObjectList() throws MyCsvReaderException;
-}
