--- 2/test-app-2/src/main/java/com/whoisacat/edu/testingApp2/service/QuizzReaderService.java
@@ -1,11 +1,0 @@
-package com.whoisacat.edu.testingApp2.service;
-
-import com.whoisacat.edu.testingApp2.dao.MyCsvReaderException;
-import com.whoisacat.edu.testingApp2.domain.Question;
-
-import java.util.List;
-
-public interface QuizzReaderService{
-
-    List<Question> readList() throws MyCsvReaderException;
-}
